+++ EquiTrack/t2f/serializers/filters/action_points.py
@@ -2,7 +2,7 @@ from __future__ import unicode_literals
 
 from rest_framework import serializers
 
-from t2f.serializers import ActionPointSerializer
+from t2f.serializers.travel import ActionPointSerializer
 from t2f.serializers.filters import SortFilterSerializer
