+++ EquiTrack/t2f/serializers/filters/travel_list.py
@@ -2,7 +2,7 @@ from __future__ import unicode_literals
 
 from rest_framework import serializers
 
-from t2f.serializers import TravelListSerializer
+from t2f.serializers.travel import TravelListSerializer
 from t2f.serializers.filters import SortFilterSerializer
